+++ battlecode22-scaffold/src/AWatchTowerBot/PathFinder.java
@@ -1,4 +1,4 @@
-package ASuperBot;
+package AWatchTowerBot;
 
 import battlecode.common.*;