+++ battlecode22-scaffold/src/OSuperBot/CombatUtil.java
@@ -1,4 +1,4 @@
-package ASuperBot;
+package OSuperBot;
 
 import battlecode.common.*;